requirements.txt
google-adk[database]==0.3.0
python-dotenv==1.1.0
deprecated